+++ app/src/main/ui/src/components/users/UserForm.tsx
@@ -1,11 +1,14 @@
-import {Button, Form, H2, ListWithSearchbar} from "@cloudogu/deprecated-ces-theme-tailwind";
+import {deprecated_Form as Form} from "@cloudogu/ces-theme-tailwind";
+import {Button, H2, ListWithSearchbar} from "@cloudogu/deprecated-ces-theme-tailwind";
 import {TrashIcon} from "@heroicons/react/24/outline";
+import {useMemo} from "react";
 import {t} from "../../helpers/i18nHelpers";
 import {useConfirmation} from "../../hooks/useConfirmation";
 import {Prompt} from "../../hooks/usePrompt";
 import useUserFormHandler from "../../hooks/useUserFormHandler";
-import {GroupsService} from "../../services/Groups";
+import { GroupsService} from "../../services/Groups";
 import {ConfirmationDialog} from "../ConfirmationDialog";
+import type {Group} from "../../services/Groups";
 import type {User} from "../../services/Users";
 import type {NotifyFunction, UseFormHandlerFunctions} from "@cloudogu/deprecated-ces-theme-tailwind";
 
@@ -29,6 +32,7 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
         notification,
         notify
     } = useUserFormHandler<T>(props.initialUser, (values: T) => props.onSubmit(values, notify, handler));
+    const isNewUser: boolean = useMemo(() => props.initialUser.username.length === 0, []);
     const {open, setOpen: toggleModal, targetName: groupName, setTargetName: setGroupName} = useConfirmation();
 
     const addGroup = (groupName: string): void => {
@@ -63,7 +67,7 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
                 exclude: handler.values.memberOf ?? [],
             }
         );
-        return groupsData.data.map(x => x.name);
+        return groupsData.data.map((x: Group) => x.name);
     };
 
     const renderGroupsList = (readonly = false, pageSize = 5) => (
@@ -96,7 +100,8 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
         <Form handler={handler}>
             {notification}
             <Form.ValidatedTextInput type={"text"} name={"username"} disabled={props.disableUsernameField ?? true}
-                data-testid="username" placeholder={t("users.placeholder.username")}>
+                data-testid="username" placeholder={t("users.placeholder.username")}
+                hint={isNewUser ? t("users.hint.username") : undefined}>
                 {t("editUser.labels.username")}
             </Form.ValidatedTextInput>
             <Form.ValidatedTextInput type={"text"} name={"givenname"} data-testid="givenname"
@@ -108,7 +113,7 @@ export default function UserForm<T extends User>(props: UserFormProps<T>) {
                 {t("editUser.labels.surname")}
             </Form.ValidatedTextInput>
             <Form.ValidatedTextInput type={"text"} name={"displayName"} data-testid="displayName"
-                placeholder={t("users.placeholder.displayName")}>
+                placeholder={t("users.placeholder.displayName")} hint={isNewUser ? t("users.hint.displayName") : undefined}>
                 {t("editUser.labels.displayName")}
             </Form.ValidatedTextInput>
             <Form.ValidatedTextInput type={"text"} name={"mail"} data-testid="mail"