+++ CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v1.13.0-1] - 2024-02-29
 ### Added
-- Add ces-theme-tailwind in a second, higher, version (v0.3.6) (#111)
-- add Cloudogu logo for better container crash visibility
+- Add ces-theme-tailwind in a second, higher, version (v0.3.7) (#111)
+- Add Cloudogu logo for better container crash visibility
+- Add descriptions for username / displayname (#114)
 
 ### Changed
 - Stream all log output to standard out, configure log level (#115)